worker/requirements.txt
certifi==2022.9.24
minio==7.1.12
nats-py==2.2.0
urllib3==1.26.13
